Commented out prune command
I tried to fix it, but it still appears that all of the program's errors (aside from the occasional UnhandledPromiseRejectionWarning) originate from that command. So for now I'm commenting it out to help less experienced users avoid the problem. If you want to use it still, simply remove the comment and the code should run